backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pymongo==4.6.1
python-dotenv==1.0.0
web3==6.15.1
requests==2.31.0
python-dateutil==2.8.2
